+++ swift/llm/utils/template.py
@@ -15,6 +15,7 @@ from transformers import PreTrainedTokenizerBase, StoppingCriteria
 from transformers.dynamic_module_utils import get_class_from_dynamic_module
 
 from swift.llm.agent.utils import calculate_loss_scale, get_tools_prompt
+from swift.llm.utils.vision_utils import load_video
 from swift.torchacc_utils import pad_and_split_batch
 from swift.utils import get_dist_setting, upper_bound, use_torchacc
 
@@ -1389,100 +1390,6 @@ class Internvl2Template(InternvlTemplate):
 
     video_segments = 8
 
-    def get_index(self, bound, fps, max_frame, first_idx=0, num_segments=32):
-        if bound:
-            start, end = bound[0], bound[1]
-        else:
-            start, end = -100000, 100000
-        start_idx = max(first_idx, round(start * fps))
-        end_idx = min(round(end * fps), max_frame)
-        seg_size = float(end_idx - start_idx) / num_segments
-        frame_indices = np.array(
-            [int(start_idx + (seg_size / 2) + np.round(seg_size * idx)) for idx in range(num_segments)])
-        return frame_indices
-
-    def build_transform(self, input_size):
-        import torchvision.transforms as T
-        from torchvision.transforms.functional import InterpolationMode
-        mean, std = (0.485, 0.456, 0.406), (0.229, 0.224, 0.225)
-        transform = T.Compose([
-            T.Lambda(lambda img: img.convert('RGB') if img.mode != 'RGB' else img),
-            T.Resize((input_size, input_size), interpolation=InterpolationMode.BICUBIC),
-            T.ToTensor(),
-            T.Normalize(mean=mean, std=std)
-        ])
-        return transform
-
-    def find_closest_aspect_ratio(self, aspect_ratio, target_ratios, width, height, image_size):
-        best_ratio_diff = float('inf')
-        best_ratio = (1, 1)
-        area = width * height
-        for ratio in target_ratios:
-            target_aspect_ratio = ratio[0] / ratio[1]
-            ratio_diff = abs(aspect_ratio - target_aspect_ratio)
-            if ratio_diff < best_ratio_diff:
-                best_ratio_diff = ratio_diff
-                best_ratio = ratio
-            elif ratio_diff == best_ratio_diff:
-                if area > 0.5 * image_size * image_size * ratio[0] * ratio[1]:
-                    best_ratio = ratio
-        return best_ratio
-
-    def dynamic_preprocess(self, image, min_num=1, max_num=6, image_size=448, use_thumbnail=False):
-        orig_width, orig_height = image.size
-        aspect_ratio = orig_width / orig_height
-
-        # calculate the existing image aspect ratio
-        target_ratios = set((i, j) for n in range(min_num, max_num + 1) for i in range(1, n + 1)
-                            for j in range(1, n + 1) if max_num >= i * j >= min_num)
-        target_ratios = sorted(target_ratios, key=lambda x: x[0] * x[1])
-
-        # find the closest aspect ratio to the target
-        target_aspect_ratio = self.find_closest_aspect_ratio(aspect_ratio, target_ratios, orig_width, orig_height,
-                                                             image_size)
-
-        # calculate the target width and height
-        target_width = image_size * target_aspect_ratio[0]
-        target_height = image_size * target_aspect_ratio[1]
-        blocks = target_aspect_ratio[0] * target_aspect_ratio[1]
-
-        # resize the image
-        resized_img = image.resize((target_width, target_height))
-        processed_images = []
-        for i in range(blocks):
-            box = ((i % (target_width // image_size)) * image_size, (i // (target_width // image_size)) * image_size,
-                   ((i %
-                     (target_width // image_size)) + 1) * image_size, ((i //
-                                                                        (target_width // image_size)) + 1) * image_size)
-            # split the image
-            split_img = resized_img.crop(box)
-            processed_images.append(split_img)
-        assert len(processed_images) == blocks
-        if use_thumbnail and len(processed_images) != 1:
-            thumbnail_img = image.resize((image_size, image_size))
-            processed_images.append(thumbnail_img)
-        return processed_images
-
-    def load_video(self, video_path, bound=None, input_size=448, max_num=1, num_segments=32):
-        from decord import VideoReader, cpu
-        from PIL import Image
-        vr = VideoReader(video_path, ctx=cpu(0), num_threads=1)
-        max_frame = len(vr) - 1
-        fps = float(vr.get_avg_fps())
-
-        pixel_values_list, num_patches_list = [], []
-        transform = self.build_transform(input_size=input_size)
-        frame_indices = self.get_index(bound, fps, max_frame, first_idx=0, num_segments=num_segments)
-        for frame_index in frame_indices:
-            img = Image.fromarray(vr[frame_index].asnumpy()).convert('RGB')
-            img = self.dynamic_preprocess(img, image_size=input_size, use_thumbnail=True, max_num=max_num)
-            pixel_values = [transform(tile) for tile in img]
-            pixel_values = torch.stack(pixel_values)
-            num_patches_list.append(pixel_values.shape[0])
-            pixel_values_list.append(pixel_values)
-        pixel_values = torch.cat(pixel_values_list)
-        return pixel_values, num_patches_list
-
     def replace_tag(self, media_type, index, example) -> List[Context]:
         if media_type == 'image':
             return [[-100]]
@@ -1532,7 +1439,7 @@ class Internvl2Template(InternvlTemplate):
             if not isinstance(videos_path, (list, tuple)):
                 videos_path = [videos_path]
             assert len(videos_path) == 1
-            pixel_values, num_patches = self.load_video(videos_path[0], num_segments=self.video_segments)
+            pixel_values, num_patches = load_video(videos_path[0], num_segments=self.video_segments)
             assert len(num_patches) == len(idx_list)
             added_tokens_len = 0
             for idx, num_patch in zip(idx_list, num_patches):
